requirements.txt
aiosqlite
selenium
webdriver-manager
python-telegram-bot[job-queue]
